feat: refactor service area form
